+++ .storybook/main.ts
@@ -7,7 +7,7 @@ const config: StorybookConfig = {
     "@storybook/addon-essentials",
     "@storybook/addon-onboarding",
     "@storybook/addon-interactions",
-      "storybook-tailwind-dark-mode"
+    "storybook-tailwind-dark-mode",
   ],
   framework: {
     name: "@storybook/react-vite",
