+++ package.json
@@ -37,6 +37,7 @@
     "chroma-js": "^2.4.2",
     "class-variance-authority": "^0.7.0",
     "jotai": "^2.7.0",
+    "lucide-react": "^0.363.0",
     "react": "^18.2.0",
     "react-aria": "^3.32.1",
     "react-dom": "^18.2.0",
