+++ src/components/Checkbox/Checkbox.stories.tsx
@@ -1,20 +1,16 @@
-import { Checkbox, CheckboxProps } from './Checkbox';
+import { Checkbox } from "./Checkbox";
 
 export default {
-  title: 'Components/Checkbox',
+  title: "Components/Checkbox",
   component: Checkbox,
   paramters: {
     controls: {
-      exclude: ['className', 'defaultChecked'],
+      exclude: ["className", "defaultChecked"],
     },
   },
 };
 
 export const Default = {
-  render: (args: CheckboxProps) => <Checkbox {...args} />,
-  args: {
-    label: 'Die Richtigkeit der Angaben wurde bestätigt',
-    value: 'Option 1',
-    disabled: true,
-  },
+  render: () => <Checkbox />,
+  args: {},
 };
