--- src/components/Checkbox/index.ts
@@ -1,1 +1,0 @@
-export { Checkbox } from './Checkbox';
