+++ src/components/index.ts
@@ -1,26 +1,27 @@
-export * from './Avatar';
-export * from './Badge';
-export * from './Button';
-export * from './Callout';
-export * from './Card';
-export * from './Checkbox';
-export * from './CopyToClipboard';
-export * from './Drawer';
-export * from './Dropdown';
-export * from './Dropdown2';
-export * from './Headings';
-export * from './Link';
-export * from './Loading';
-export * from './Navigation';
-export * from './MenuItem';
-export * from './ModalDialog';
-export * from './Layout';
-export * from './RadioButton';
-export * from './Select';
-export * from './Shimmer';
-export * from './SideNav';
-export * from './Switch';
-export * from './Tab';
-export * from './Table';
-export * from './TextField';
-export * from './Tooltip';
+export * from "./Avatar";
+export * from "./Badge";
+export * from "./Button";
+export * from "./Callout";
+export * from "./Card";
+export * from "./CopyToClipboard";
+export * from "./Drawer";
+export * from "./Dropdown";
+export * from "./Dropdown2";
+export * from "./Headings";
+export * from "./Link";
+export * from "./Loading";
+export * from "./Navigation";
+export * from "./MenuItem";
+export * from "./ModalDialog";
+export * from "./Layout";
+export * from "./RadioButton";
+export * from "./Select";
+export * from "./Shimmer";
+export * from "./SideNav";
+export * from "./Switch";
+export * from "./Tab";
+export * from "./Table";
+export * from "./TextField";
+export * from "./Tooltip";
+
+export * from "./../legacy/Checkbox";
