+++ src/legacy/Checkbox/Checkbox.stories.tsx
@@ -1,0 +1,20 @@
+import { Checkbox, CheckboxProps } from "./Checkbox";
+
+export default {
+  title: "Legacy/Components/Checkbox",
+  component: Checkbox,
+  paramters: {
+    controls: {
+      exclude: ["className", "defaultChecked"],
+    },
+  },
+};
+
+export const Default = {
+  render: (args: CheckboxProps) => <Checkbox {...args} />,
+  args: {
+    label: "Die Richtigkeit der Angaben wurde bestätigt",
+    value: "Option 1",
+    disabled: true,
+  },
+};
