+++ src/legacy/Checkbox/index.ts
@@ -1,0 +1,1 @@
+export { Checkbox as LegacyCheckbox } from "./Checkbox";
